ISC-16617 Sitewide style fixes/tweaks
For <hr>: Created two classes .hr-max and .hr-min with top and bottom margins 18px and 5px respectively. 

For the first H2: Added CSS so the first H2 that comes after the H1 has a top margin of 0.

For underline wiggle: Added a fix to the CSS, it inherited it from the UW-2014 theme and had to do with the transition.

For a.uw-btn.btn-sm: Added CSS which gives it a top margin of 10px.